pyproject.toml
[project]
name = "simple-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "llama-cloud-services>=0.6.81",
    "llama-parse>=0.6.81",
    "ollama>=0.6.1",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
